move all methods to mixins.py
